bot/requirements.txt
anyio==4.6.0
APScheduler==3.6.3
bcrypt==4.2.0
cachetools==4.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
paramiko==3.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tzlocal==5.2
urllib3==2.2.3
